fix(xcm): enable proper handling of XCMP messages
XCMP messages were not reaching the destination sibling
chain due to missing configuration of the `ParachainSystem`
message queues.
