+++ runtime/frequency/src/xcm_queue.rs
@@ -1,8 +1,6 @@
-// use crate::{MessageQueue, ParachainSystem, RuntimeBlockWeights, RuntimeCall, RuntimeEvent};
-// use crate::{ParachainSystem, RuntimeBlockWeights, RuntimeCall, RuntimeEvent, AccountId, RuntimeOrigin, Runtime};
 use crate::{
 	AccountId, MessageQueue, ParachainSystem, Runtime, RuntimeBlockWeights, RuntimeCall,
-	RuntimeEvent, RuntimeOrigin, XcmpQueue,
+	RuntimeEvent, XcmpQueue,
 };
 
 use frame_support::{
@@ -11,18 +9,13 @@ use frame_support::{
 	weights::Weight,
 };
 
-use polkadot_parachain_primitives::primitives::Sibling;
+use crate::xcm_commons::XcmOriginToTransactDispatchOrigin;
 
 use frame_system::EnsureRoot;
 
 use cumulus_primitives_core::{AggregateMessageOrigin, ParaId};
 
 use parachains_common::message_queue::{NarrowOriginToSibling, ParaIdToSibling};
-
-use staging_xcm_builder::{
-	AccountId32Aliases, ParentIsPreset, RelayChainAsNative, SiblingParachainAsNative,
-	SiblingParachainConvertsVia, SignedAccountId32AsNative, SovereignSignedViaLocation,
-};
 
 // use xcm_config;
 #[cfg(not(feature = "runtime-benchmarks"))]
@@ -32,10 +25,6 @@ use polkadot_runtime_common::xcm_sender::NoPriceForMessageDelivery;
 
 pub use sp_runtime::Perbill;
 
-use staging_xcm::latest::prelude::*;
-
-use pallet_xcm::XcmPassthrough;
-
 #[cfg(not(feature = "runtime-benchmarks"))]
 use xcm_executor;
 
@@ -43,50 +32,6 @@ parameter_types! {
 	pub MessageQueueServiceWeight: Weight = Perbill::from_percent(35) * RuntimeBlockWeights::get().max_block;
 }
 
-parameter_types! {
-	// update for dev
-	pub const RelayNetwork: Option<NetworkId> = Some(NetworkId::Polkadot);
-	pub RelayChainOrigin: RuntimeOrigin = cumulus_pallet_xcm::Origin::Relay.into();
-}
-
-/// Type for specifying how a `Location` can be converted into an `AccountId`. This is used
-/// when determining ownership of accounts for asset transacting and when attempting to use XCM
-/// `Transact` in order to determine the dispatch Origin.
-/// Conversions between Multilocation to an accountid
-/// Parachain origin is converted to corresponding souverin account
-pub type LocationToAccountId = (
-	// The parent (Relay-chain) origin converts to the parent `AccountId`.
-	ParentIsPreset<AccountId>,
-	// Sibling parachain origins convert to AccountId via the `ParaId::into`.
-	SiblingParachainConvertsVia<Sibling, AccountId>,
-	// Straight up local `AccountId32` origins just alias directly to `AccountId`.
-	AccountId32Aliases<RelayNetwork, AccountId>,
-);
-
-///////////
-
-/// This is the type we use to convert an (incoming) XCM origin into a local `Origin` instance,
-/// ready for dispatching a transaction with Xcm's `Transact`. There is an `OriginKind` which can
-/// biases the kind of local `Origin` it will become.
-pub type XcmOriginToTransactDispatchOrigin = (
-	// Sovereign account converter; this attempts to derive an `AccountId` from the origin location
-	// using `LocationToAccountId` and then turn that into the usual `Signed` origin. Useful for
-	// foreign chains who want to have a local sovereign account on this chain which they control.
-	SovereignSignedViaLocation<LocationToAccountId, RuntimeOrigin>,
-	// Native converter for Relay-chain (Parent) location; will convert to a `Relay` origin when
-	// recognized.
-	RelayChainAsNative<RelayChainOrigin, RuntimeOrigin>,
-	// Native converter for sibling Parachains; will convert to a `SiblingPara` origin when
-	// recognized.
-	SiblingParachainAsNative<cumulus_pallet_xcm::Origin, RuntimeOrigin>,
-	// Native signed account converter; this just converts an `AccountId32` origin into a normal
-	// `RuntimeOrigin::Signed` origin of the same 32-byte value.
-	SignedAccountId32AsNative<RelayNetwork, RuntimeOrigin>,
-	// Xcm origins can be represented natively under the Xcm pallet's Xcm origin.
-	XcmPassthrough<RuntimeOrigin>,
-);
-
-//////////////////////////
 impl cumulus_pallet_xcmp_queue::Config for Runtime {
 	type RuntimeEvent = RuntimeEvent;
 	type ChannelInfo = ParachainSystem;
